+++ src/interfaces/Hypotheses.java
@@ -4,8 +4,8 @@ import basic_hierarchy.interfaces.Hierarchy;
 
 public interface Hypotheses {
 	public void calculate(Hierarchy h);
-	public int getTP();
-	public int getFP();
-	public int getTN();
-	public int getFN();
+	public long getTP();
+	public long getFP();
+	public long getTN();
+	public long getFN();
 }
